+++ attributes/default.rb
@@ -21,9 +21,9 @@
 #
 
 # Download data
-default['openresty']['source']['version']   = '1.4.3.9'
+default['openresty']['source']['version']   = '1.5.8.1'
 default['openresty']['source']['url']       = "http://agentzh.org/misc/nginx/ngx_openresty-#{node['openresty']['source']['version']}.tar.gz"
-default['openresty']['source']['checksum']  = '353131c73bc1b2ab7c55a3f6982d293bb19c2299122b1ac04a614141f5b271fa'
+default['openresty']['source']['checksum']  = '96ec68efb1ecd83a3b4871d1051505d858d13cb866f6be644b8f60ed9397fbc2'
 
 # Directories
 default['openresty']['dir']                 = '/etc/nginx'